move all misc tests to test_misc.py
